remove warning

+++ 05-SceneManager/GameObject.h
@@ -25,7 +25,7 @@ protected:
 	float vx;
 	float vy;
 
-	int nx;	 
+	float nx;	 
 
 	int state;
